+++ tests/specs/references/typeOperator/readonly.d.ts
@@ -1,0 +1,1 @@
+declare function test(array: readonly string[])